Fixed download status message so that it prints the % in integers.

+++ src/net/orpiske/sdm/lib/net/Downloader.java
@@ -66,14 +66,15 @@ public class Downloader {
 				else {
 					percentComplete = 0.0;
 				}
-					
-				System.out.print("\r" + percentComplete + "% complete (" + i + " of " 
+				
+			
+				System.out.print("\r" + (int) percentComplete + "% complete (" + i + " of " 
 						+ total + ")");
 			}
 		}
 		
 		System.out.print("\r100% complete (" + i + " of " + total + ")\n");
-
+		
 		output.flush();
 
 	}
